finished prototyping roster menu added new game button, debug actions to randomize free agents and change difficulty

+++ src/MtM/model/business/service/MissionGenerator.java
@@ -48,7 +48,7 @@ public class MissionGenerator {
         ret = new Mission(type, 100, 1);
         ret.setPrimaryStat(primaryStat);
         ret.setSecondaryStat(secondaryStat);
-        ret.setReward(144 * difficulty * difficulty);
+        ret.setReward((int)(144.0 * difficulty * difficulty));
 
         return ret;
     }
